Convert sparse matrix to a dense one
I encountered a ValueError (shape mismatch) following the tutorial on running `python src/main.py --config configs/causal_gan.cfg --create_grn`. On investigating the error, it seems the data was read as sparse which pandas dataframe expects dense matrix. 

We can fix this in two ways: using `.todense()` or `.toarray()` depending on whether we want to return a `numpy.matrix` or `numpy.ndarray` object. Based on my research, seems the later is preferred as the use of `numpy.matrix` may be deprecated in the future. 

I tested this fix and I can successfully run the `--create_grn` function.